Add a class for computing checksums, counting bytes and reporting progress
At the moment, the storage backend is responsible for computing checksums and
reporting the size of files based on the number of bytes read. However, these
are common concerns across all storage backends, and so this functionality can
be pulled the other side of the storage backend interface.

This class is therefore intended to wrap any file-like object, and to checksum
the file as it is read. It also counts bytes and calls the progress callback if
one is provided.

In a subsequent commit we'll integrate it with FileInstance, and remove the
functionality from the PyFS storage backend.